+++ src/modules/index.ts
@@ -57,7 +57,7 @@ function getSetup(config: RhinestoneAccountConfig): ModeleSetup {
     validators.push(socialRecoveryValidator)
   }
 
-  const intentExecutorAddress = config.useDev
+  const intentExecutorAddress = config.orchestratorUrl
     ? INTENT_EXECUTOR_ADDRESS_DEV
     : INTENT_EXECUTOR_ADDRESS
   const executors: Module[] = [