+++ src/orchestrator/types.ts
@@ -70,7 +70,7 @@ interface Execution {
   data: Hex
 }
 
-type SettlementLayer = 'SAME_CHAIN' | 'ACROSS' | 'ECO'
+export type SettlementLayer = 'SAME_CHAIN' | 'ACROSS' | 'ECO' | 'RELAY'
 
 interface IntentOptions {
   topupCompact: boolean
@@ -126,10 +126,10 @@ interface IntentInput {
     amount?: bigint
   }[]
   accountAccessList?: AccountAccessList
-  options?: IntentOptions
+  options: IntentOptions
 }
 
-type SettlementSystem = 'SAME_CHAIN' | 'ACROSS'
+type SettlementSystem = 'SAME_CHAIN' | 'ACROSS' | 'RELAY'
 
 interface IntentCost {
   hasFulfilledAll: boolean